requirements.txt
Django==3.2.9
djangorestframework==3.12.4
drf-yasg==1.20.0
Pillow
psycopg2>=2.8